feat: add katana item with tail-cutting functionality
Introduce a new katana item that allows players to cut enemy tails and convert them to apples. The item has a spawn weight of 15 and a duration of 1. The `do` method in the `Player` class is updated to pass the `game` instance for apple spawning. The katana item is added to the spawner's item classes list.